fix(BRI-60) | изменил ссылки на OAuth
