+++ packages/client/src/pages/GameStart/GameStart.tsx
@@ -1,9 +1,30 @@
-import { Link } from 'react-router-dom'
+import { Link, useNavigate } from 'react-router-dom'
 import { GameLayout } from '../../layouts/GameLayout'
 import styles from './GameStart.module.scss'
 import logoutIcon from '/images/icons/logout-icon.svg'
+import { useEffect, useState } from 'react'
+import { useAppDispatch, useAppSelector } from '../../store'
+import { sigInYandex } from '../../store/auth/authSlice'
 
 export function GameStart(): JSX.Element {
+  const [isMounted, setIsMounted] = useState(true)
+  const dispatch = useAppDispatch();
+  const { verificate } = useAppSelector(state => state.auth)
+  const navigate = useNavigate()
+  
+  useEffect(() => {
+    if(verificate && !isMounted) {
+      dispatch(sigInYandex(
+        {code: String(new URL(window.location.href).searchParams.get('code')), redirect_uri: 'https://quiz-game-client.vercel.app/start'}
+    )).then((response) => {
+        if(response.payload == 'Произошла ошибка') {
+            navigate('/auth')
+        }
+    })
+    }
+    setIsMounted(false)
+  }, [isMounted])
+  
   return (
     <GameLayout>
       <>
